pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "montaris-x"
version = "0.1.0"
description = "Cross-platform ROI editor for scientific images"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "numpy>=1.24.0",
    "tifffile>=2023.7.0",
    "Pillow>=10.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
]

[project.scripts]
montaris = "montaris.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
qt_api = "pyside6"
